Added IFragmentEvents interface connecting fragments.
First of all - this is a poor way to implement this. Must be refactored.
Currently the IFragmentEvents is implemented in BaseFragment that
supplies default impl. for each event call. The default impl. is just
doing *nothing*. The home and invoice fragments are connected to each
other via the IFragmentEvents when created in MainActivity. Each
Fragment decides to which event from the interface to react.