Changed default install location to 'auto'.
The application may be installed on the external storage, but the system
will install the application on the internal storage by default. 

Until 3.1.0 the app installed itself automatically on the external
storage. Thus, the MyGlob widgets were not available. Widgets may only
be added when the app runs on the internalStorage.

Bumped version to 3.1.1.